Moves 'instructions' from to-do.js to new file instructions.js
